Move example hosts/tasks to fixtures file, fix spelling of fixtures filename
